@@ -1,2 +1,3 @@
 // if the user does not specify a theme, we still need our alias to point somewhere
-module.exports = {};
+export default {};
+
